Add files via upload
Added Troup Fits. Using the method of finding R, L, and M using the Troup paper and comparing it to the isochrone fitting. 
Changed the name of Bianry_Data_Clean2 to BinaryDataClean2 to be in line with PEP standards for class naming. 
Make some changes to Binary_Fraction to work with the naming convention for the Troup fits.
BDC also does error analysis for the troup fits