Add files via upload
We are now on BinaryDataClean3. Changed the way we do the isochrone fitting. Now it looks for exact matches and slowly expands the range of possible fits. Most of the time this makes it so that we only find one isochrone fit, sometimes we find 2 or 3, and we average over those, but this works rather well. 

Binary_Fraction.py I think is mostly unchanged. I changed the normal distribution paramiters. I need to make mean and std inputs to the period distribution function. There's also a few hacky bits where I have to go in and change a few lines depending on if I want to use the Troup fits or the Isochrone fits. 

New Isochrone Method notebook, is the same old notebook that i've been running things with, but using BDC3, and a slightly different log normal distribution, that's why the synthetic lines have a much shallower slope